refactor(fieldmap): remove data fetch service
