+++ js/source/include/fieldmap/hooks/useDownloadPlotOrder.ts
@@ -1,0 +1,37 @@
+import { useLayoutConfig } from '../contexts/LayoutConfigContext';
+import { useView } from '../contexts/ViewContext';
+import { DownloadOpts } from '../model.types';
+
+export const useDownloadPlotOrder = () => {
+	const {
+		activeTrialIds
+	} = useView();
+
+	const {
+		topBorder,
+		rightBorder,
+		bottomBorder,
+		leftBorder
+	} = useLayoutConfig();
+
+	const downloadPlotOrder = (downloadOpts: DownloadOpts) => {
+		const q = new URLSearchParams({
+			trial_ids: activeTrialIds.join(','),
+			type: downloadOpts.type,
+			order: downloadOpts.order,
+			start: downloadOpts.start,
+			top_border: String(downloadOpts.borders && topBorder),
+			right_border: String(downloadOpts.borders && rightBorder),
+			bottom_border: String(downloadOpts.borders && bottomBorder),
+			left_border: String(downloadOpts.borders && leftBorder),
+			gaps: String(downloadOpts.gaps),
+			subplots: String(downloadOpts.subplots),
+			plants: String(downloadOpts.plants),
+			hm_pltid: downloadOpts.hmPltid,
+			hm_range: downloadOpts.hmRange,
+			hm_row: downloadOpts.hmRow
+		}).toString();
+		window.open(`/ajax/breeders/trial_plot_order?${q}`, '_blank');
+	};
+	return { downloadPlotOrder };
+};
