+++ js/source/include/fieldmap/hooks/useSubmitGeoLayout.ts
@@ -1,0 +1,30 @@
+import { useModals } from '../contexts/ModalsContext';
+import { usePlotGrid } from '../contexts/PlotGridContext';
+
+export const useSubmitGeoLayout = () => {
+	const {
+		setLoading
+	} = useModals();
+
+	const {
+		fetchObservationUnits
+	} = usePlotGrid();
+
+    const submitGeoLayout = async () => {
+        const fm = (window as any).geoFieldMapInstance;
+        if (fm) {
+            setLoading(true);
+			try {
+				const msg = await fm.update();
+				alert(msg || 'Geo layout updated successfully!');
+				fetchObservationUnits();
+			} catch (e) {
+				alert(e || 'Failed to update geo layout');
+			} finally {
+				setLoading(false);
+			}
+        }
+    };
+
+	return { submitGeoLayout };
+};
